Bucket Size 4 (#136)
Bench: 9314841

ELO | 0.24 +- 2.81 (95%)
SPRT | 8.0+0.08s Threads=1 Hash=16MB
LLR | 2.98 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 24120 W: 4972 L: 4955 D: 14193

Just more entries.

+++ src/move.h
@@ -29,18 +29,18 @@ extern const char* SQ_TO_COORD[];
 #define BuildMove(start, end, piece, promo, cap, dub, ep, castle)                                                      \
   (start) | ((end) << 6) | ((piece) << 12) | ((promo) << 16) | ((cap) << 20) | ((dub) << 21) | ((ep) << 22) |          \
       ((castle) << 23)
-#define MoveStart(move) ((move)&0x3f)
-#define MoveEnd(move) (((move)&0xfc0) >> 6)
-#define MovePiece(move) (((move)&0xf000) >> 12)
-#define MovePromo(move) (((move)&0xf0000) >> 16)
-#define MoveCapture(move) (((move)&0x100000) >> 20)
-#define MoveDoublePush(move) (((move)&0x200000) >> 21)
-#define MoveEP(move) (((move)&0x400000) >> 22)
-#define MoveCastle(move) (((move)&0x800000) >> 23)
+#define MoveStart(move) ((int)(move)&0x3f)
+#define MoveEnd(move) (((int)(move)&0xfc0) >> 6)
+#define MovePiece(move) (((int)(move)&0xf000) >> 12)
+#define MovePromo(move) (((int)(move)&0xf0000) >> 16)
+#define MoveCapture(move) (((int)(move)&0x100000) >> 20)
+#define MoveDoublePush(move) (((int)(move)&0x200000) >> 21)
+#define MoveEP(move) (((int)(move)&0x400000) >> 22)
+#define MoveCastle(move) (((int)(move)&0x800000) >> 23)
 // just mask the start/end bits into a single int for indexing butterfly tables
-#define MoveStartEnd(move) ((move)&0xfff)
+#define MoveStartEnd(move) ((int)(move)&0xfff)
 
-#define Tactical(move) (((move)&0x1f0000) >> 16)
+#define Tactical(move) (((int)(move)&0x1f0000) >> 16)
 
 Move ParseMove(char* moveStr, Board* board);
 char* MoveToStr(Move move);
